requirements.txt
tensorflow_gpu==1.8.0
numpy==1.14.3
opencv_python==3.2.0.7
